Change the flow of the game using JS and the DOM.
Create buttons to use for game input instead of a prompt.
Use JS to add a new div for p elements for keeping score.
Use click event listeners to keep track of player selection.
Create a button to restart the game once someone gets to 5 wins.
Add a new image to the top of the page.

Unfortunately, I wasn't committing enough during the process since I was into getting the code correct.
It is now a more functional version of Rock, Paper, Scissors.